new upgrade plan design
